:sparkles: Add support for extensions in destination paths and schema conversion
This commit adds support for the `extensions` folder in the `DestinationPaths` class. It also updates the `SchemaSource` enum to include a new value for extensions. The `Converter` class has been modified to generate schemas from the `extensions` folder as well. Additionally, the `InterfaceBuilder` class now generates interfaces for extension schemas and handles attribute modifications specific to extensions.